+++ TrendByPivotPointsOptimizator/SecurityData.cs
@@ -6,7 +6,7 @@ namespace TrendByPivotPointsOptimizator
     {
         public string Name;
         public Currency Currency;
-        public int Shares;
+        public double Shares;
         public double CommissionRate;
         public bool IsUSD;
         public double RateUSD;
