+++ TrendByPivotPointsOptimizator/Ticker.cs
@@ -8,14 +8,14 @@ namespace TrendByPivotPointsOptimizator
         public List<Bar> Bars { get; set; }
         public string Name { get; private set; }
         public Currency Currency { get; private set; }
-        public int Shares { get; private set; }
+        public double Shares { get; private set; }
         public double CommissionRate { get; private set; }
         public bool IsUSD { get; private set; }
         public double RateUSD { get; private set; }
         public Logger Logger { get; private set; }
         public List<Bar> InitBars { get; }
 
-        public Ticker(string name, Currency currency, int shares, List<Bar> bars, Logger logger, 
+        public Ticker(string name, Currency currency, double shares, List<Bar> bars, Logger logger, 
             double commissionRate, bool isUSD, double rateUSD)
         {
             Name = name;
